Add everything into one model folder
