+++ server/gql/models/studentSubjects/index.js
@@ -1,31 +1,12 @@
-import { studentSubject } from './query';
+import { GraphQLStudentSubject } from './model';
 import { studentSubjectConnection } from './list';
-import { GraphQLInt, GraphQLNonNull } from 'graphql';
-import db from '@server/database/models';
+import { StudentSubjectQueries } from './query';
 import { studentSubjectMutation } from './mutation';
 
-export const StudentSubject = studentSubject;
+export const StudentSubject = GraphQLStudentSubject;
 
-// export const StudentSubjectConnection = studentSubjectConnection;
 export const StudentSubjectConnection = studentSubjectConnection;
 
-// queries on the studentSubjects table
-export const studentSubjectQueries = {
-  args: {
-    id: {
-      type: new GraphQLNonNull(GraphQLInt)
-    }
-  },
-  query: {
-    type: studentSubject
-  },
-  list: {
-    ...StudentSubjectConnection,
-    resolve: StudentSubjectConnection.resolve,
-    type: StudentSubjectConnection.connectionType,
-    args: StudentSubjectConnection.connectionArgs
-  },
-  model: db.studentSubjects
-};
+export const studentSubjectQueries = StudentSubjectQueries;
 
 export const studentSubjectMutations = studentSubjectMutation;
