+++ server/gql/models/subjects/tests/subjects.test.js
@@ -3,7 +3,7 @@ import { QueryRoot } from '../../../queries';
 import { MutationRoot } from '../../../mutations';
 import { GraphQLSchema, graphqlSync } from 'graphql';
 import { createFieldsWithType, expectSameTypeNameOrKind } from '@server/utils/testUtils';
-import { subjectFields } from '../../subjects';
+import { subjectFields } from '..';
 import { timestamps } from '@gqlFields/timestamps';
 
 const schema = new GraphQLSchema({ query: QueryRoot, mutation: MutationRoot });
